fix layout issues
