+++ src/app/(business)/layout.tsx
@@ -1,44 +1,19 @@
-import { Crimson_Text, Roboto } from 'next/font/google';
-import "../globals.css";
-import Providers from "../providers";
 import { AppSidebar } from "@/components/AppSidebar";
 import { TopBar } from "@/components/top-bar";
 
-const crimsonText = Crimson_Text({
-  weight: ['400', '600', '700'],
-  style: ['normal', 'italic'],
-  subsets: ['latin'],
-  variable: '--font-crimson',
-});
-
-const roboto = Roboto({
-  subsets: ['latin'],
-  weight: ['400', '700'],
-  style: ['normal', 'italic'],
-  display: 'swap',
-});
-
-export default function RootLayout({
+export default function BusinessLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={`${crimsonText.variable} ${roboto.className} antialiased`}>
-        <Providers>
-          <div className="flex h-screen bg-background">
-            <AppSidebar />
-            <div className="flex flex-col flex-1 overflow-hidden">
-              <TopBar />
-              <main className="flex-1 overflow-x-hidden overflow-y-auto p-6">
-                {children}
-              </main>
-            </div>
-          </div>
-        </Providers>
-      </body>
-    </html>
+    <div className="flex h-screen bg-background">
+      <AppSidebar />
+      <div className="flex flex-col flex-1 overflow-hidden">
+        <TopBar />
+        <main className="flex-1 overflow-x-hidden overflow-y-auto p-6">{children}</main>
+      </div>
+    </div>
   );
 }
